kebab naming convension adopted to files not react related; TaskList mocks tests moved to mocks file
