+++ src/components/TasksList/TasksList.tsx
@@ -7,8 +7,8 @@ import AvatarHeader from 'components/AvatarHeader';
 import NavBar from 'components/NavBar';
 import { useQuery, useMutation } from '@apollo/react-hooks';
 import { QueryNewbieArgs, Query, Task, Mutation } from 'buddy-app-schema';
-import TASK_LIST from 'graphql/taskList.graphql';
-import UPDATE_TASK_STATUS from 'graphql/updateTaskStatus.graphql';
+import TASK_LIST from 'graphql/task-list.graphql';
+import UPDATE_TASK_STATUS from 'graphql/update-task-status.graphql';
 import { useParams } from 'react-router-dom';
 import Box from '@material-ui/core/Box';
 import TaskTabsContent from 'components/TaskTabsContent';
