+++ src/components/TasksList/__tests__/TasksList.test.tsx
@@ -4,7 +4,7 @@ import waitForExpect from 'wait-for-expect';
 
 import { MemoryRouter, Route } from 'react-router';
 import { MockedProvider } from '@apollo/react-testing';
-import TASK_LIST from 'graphql/taskList.graphql';
+import { taskListResponse } from '__mocks__';
 import TasksList from '../TasksList';
 
 jest.mock('@material-ui/core/AppBar', () => 'AppBar');
@@ -21,54 +21,46 @@ describe('Component - TasksList', () => {
   const path = '/buddy/newbies/1234/tasks';
 
   describe('When there are tasks', () => {
-    const taskListWithDataResponse = [
-      {
-        request: {
-          query: TASK_LIST,
-          variables: {
-            newbieId: '1234',
-          },
+    const variables = {
+      newbieId: '1234',
+    };
+
+    const newbieMockData = {
+      buddyTasks: [
+        {
+          id: '1',
+          description: 'New task description',
+          status: 'UNCOMPLETED',
+          title: 'New task title ',
         },
-        result: {
-          data: {
-            newbie: {
-              buddyTasks: [
-                {
-                  id: '1',
-                  description: 'New task description',
-                  status: 'UNCOMPLETED',
-                  title: 'New task title ',
-                },
-                {
-                  id: '2',
-                  description: 'New task description 2',
-                  status: 'COMPLETED',
-                  title: 'New task title 2',
-                },
-              ],
-              newbieTasks: [
-                {
-                  id: '3',
-                  description: 'New task description',
-                  status: 'COMPLETED',
-                  title: 'New task title ',
-                },
-                {
-                  id: '4',
-                  description: 'New task description 2',
-                  status: 'COMPLETED',
-                  title: 'New task title 2',
-                },
-              ],
-            },
-          },
+        {
+          id: '2',
+          description: 'New task description 2',
+          status: 'COMPLETED',
+          title: 'New task title 2',
         },
-      },
-    ];
+      ],
+      newbieTasks: [
+        {
+          id: '3',
+          description: 'New task description',
+          status: 'COMPLETED',
+          title: 'New task title ',
+        },
+        {
+          id: '4',
+          description: 'New task description 2',
+          status: 'COMPLETED',
+          title: 'New task title 2',
+        },
+      ],
+    };
 
     test('renders correctly', async () => {
       const component = create(
-        <MockedProvider mocks={taskListWithDataResponse} addTypename={false}>
+        <MockedProvider
+          mocks={taskListResponse(variables, newbieMockData)}
+          addTypename={false}>
           <MemoryRouter initialEntries={[path]}>
             <Route path={'/buddy/newbies/:newbieId/tasks'}>
               <TasksList />
@@ -88,27 +80,20 @@ describe('Component - TasksList', () => {
   });
 
   describe('when there are not tasks', () => {
-    const taskListEmptyStateResponse = [
-      {
-        request: {
-          query: TASK_LIST,
-          variables: {
-            newbieId: '1234',
-          },
-        },
-        result: {
-          data: {
-            newbie: {
-              buddyTasks: [],
-              newbieTasks: [],
-            },
-          },
-        },
-      },
-    ];
+    const variables = {
+      newbieId: '1234',
+    };
+
+    const newbieMockData = {
+      buddyTasks: [],
+      newbieTasks: [],
+    };
+
     test('should render empty state', async () => {
       const component = create(
-        <MockedProvider mocks={taskListEmptyStateResponse} addTypename={false}>
+        <MockedProvider
+          mocks={taskListResponse(variables, newbieMockData)}
+          addTypename={false}>
           <MemoryRouter initialEntries={[path]}>
             <Route path={'/buddy/newbies/:newbieId/tasks'}>
               <TasksList />
