+++ src/graphql/newbie-select.graphql.ts
